funciones mqtt

+++ src/string_management.h
@@ -14,7 +14,7 @@
 
 #include "driver_bg96.h"
 
-em_bg96_error_handling send_data(char *cmd,char*expect,char*buffer,uint16_t timeout);
+em_bg96_error_handling send_data(char *cmd,char*expect,char*buffer,uint32_t timeout);
 
 void parse_string(char *data,char firts_caracter,char second_caracter,char *buffer);
